feat(vm): add dynamic programming interface
Add the ability to dynamically assign address to the
source code without specifying the destination address
(numerical alias name).

This is achieved by changing the behavior of "s" and
"LABLE" alias and with the addition of jump instructions
which use stack head as jump destination.

It also changes the order in which the source cfg files
should be executed, main should always be executed first
followed by the dependencies and function libraries. In
order to avoid this issue "origin" alias should be used
in the cfg to specify which position the code following
it should go.