fix(vm): fix jif_vfs wrong vector position bug
Fix bug introduced in jif_vfs in commit 97d4954.

jif_vfs uses value below stack head (sp - 2) to
determine if to branch or not, and the stack
head (sp - 1) as the branch address. In commit
97d4954, the above mentioned behaviour was
reversed, resulting in unpredictable behaviour.
In most of the cases it would branch to address 0.