Fix conditional branch vector/value from stack bug
this commit fixes bug introduced in jif_vfs in commit 8c9d180.

Details:
jif_vfs uses value below stack head (stack.pointer - 2) to determine
wether to branch or not, and the stack head as the (stack.pointer -1)
as the branch address. In commit 8c9d180, the above mentioned behaviour
was reversed, resulting in unpredictable behaviour. In most of the cases
it would branch to address 0.